+++ src/cqterrain/tile/__init__.py
@@ -8,3 +8,4 @@ from .chamferFrame import chamfer_frame
 from .rivet import rivet
 from .slot import slot
 from .slotDiagonal import slot_diagonal
+from .glyph import glyph